Switched baud to specify actual baud, enabled interrupt driven transmission
Baud used to be specified in kilobaud but now is specified in actual baud so that you can
set odd baud rates like 33,333 baud.

There had already been preliminary code for interrupt driven TX but now it is complete. The sendFrame function will send a frame if there is an open mailbox or queue it if not.