Enhancement to setNumTXBoxes()
Alter setNumTXBoxes() to return the actual number of Tx boxes that were set.  Allows for application to discover the physical quantity of hardware mailboxes, ala:

  int MB_Qty = setNumTXBoxes(999);                               // How many Mailboxes are there?
      setNumTXBoxes(MB_qty/2);                                       // Configure 1/2 of them as Tx boxes, we can use the rest for Rx filtering.
   . . . . . .



Then the application can recall setNumTXBoxes() with the actual desired quantity.  As this change is from void to int, it should cause no issues with existing applications.